misc: add validate for valid product, add warning when click unverified product
